requirements.txt
gunicorn==20.0.4
dynaconf==2.2.3
uvicorn==0.11.3
celery==4.4.2
aiopg==1.0.0
SQLAlchemy==1.3.16
httpx==0.16.1
marshmallow-sqlalchemy==0.22.3
requests==2.23.0
Werkzeug==1.0.1
ujson==2.0.3
aiohttp==3.6.2
pika==1.1.0
redis==3.4.1
jinja2>=2.0
PyYaml>=5.0
databases[postgresql]
aioredis==1.3.1
pydicom==2.0.0
redlock==1.2.0
aioredlock==0.3.0
pyjwt==1.7.1
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.2.0
python-keycloak==0.24.0
reportlab==3.5.45
opentelemetry-api==0.14b0
opentelemetry-sdk==0.14b0
opentelemetry-instrumentation-asgi==0.14b0
opentelemetry-exporter-jaeger==0.14b0
opentelemetry-instrumentation-celery==0.14b0
opentelemetry-instrumentation-requests==0.14b0
opentelemetry-instrumentation-redis==0.14b0
fastapi==0.62.0
python-multipart
python-jose[cryptography]
elasticsearch==7.9.1
loguru==0.5.3
beren==0.7.0
